整理完毕一遍代码。

+++ src/constant.cc
@@ -2,17 +2,17 @@
 
 namespace mynum {
 
-Numeric __pi = "3.141592653589793238462643383279502884197169399375105820974944";
-Numeric __half_pi = "1.570796326794896619231321691639751442098584699687552910487472";
-Numeric __e = "2.7182818284590452353602874713527";
+Float __pi = "3.141592653589793238462643383279502884197169399375105820974944";
+Float __half_pi = "1.570796326794896619231321691639751442098584699687552910487472";
+Float __e = "2.7182818284590452353602874713527";
 
-Numeric approximate_pi(const Numeric& precision) {
-  Numeric res;
+Float approximate_pi(const Float& precision) {
+  Float res;
   return res;
 }
 
-Numeric approximate_e(const Numeric& precision) {
-  Numeric res;
+Float approximate_e(const Float& precision) {
+  Float res;
   return res;
 }
