+++ src/float/is_none.cc
@@ -1,8 +1,8 @@
-#include <mynum/integer.h>
+#include <mynum/float.h>
 
 namespace mynum {
 
-bool is_none(const Integer& num1) {
+bool is_none(const Float& num1) {
   if (num1.infinite()) return false;
   return (num1.integer_park().empty() || 
           num1.decimal_park().empty()) &&
